Extract x/y to Location
Removing a data clump. Mostly.

Still digging in and grabbing the x/y values of the Location in a couple
of spots, but I expect that will go away later.

Also trying to improve some names

add_cell -> add_living, to clarify what kind of cell is being added.

the tempvar of living_neighbor_count is renamed to avoid confusion
with the existing method name.