Extract world
Trying a different approach here. Last time
(d7230bcf70ffb92a46f5aea351b6fe758428f754) I tried to extract just Rows
and found it painful.

This time I'm trying to extract a World, mostly with an eye on
simplifying `living_neighbors_of`

In this approach, the `Conway` struct is home to the `tick` logic and
how to output the world to a string.

The `World` is a collection of `Location`s, each of which holds a
`Cell`. It also holds a private state, which our good-old nested Vector.

Location gets some useful logic here, being able to return the neighbors
of itself. Current approach is a bit brute-forcey, but it's not like
it's ever going to need to change

With this, a world can return the valid neighbors of any location, which
e can then use to determine the number of living cells neighboring any
other cell.

This works better than just trying to extract Rows. I think the Row
concept wasn't a full-fledged thing in the system, which is what made it
so hard to extract.