Non-working extraction of (location, cell)
The idea here was that locations and cells could be thrown together in a
Vec of tuples. Seemed like a reasonable way of smoothing out the weird
edges of World (i.e., `add_location` and `cell_at`). If I just had a
iterable collection of tuples then I could iterate thorugh them without
worrying about location or if it contained a cell.

Problem is the ordering. If `world` creates its tuple-vec at creation,
then `add_cell` has to swap out dead for living cells in the middle of
the vec. Which still requires some logic to see whether or not that's
allowed.

Which doesn't seem much better than what I'm currently doing, honestly.

The other option (that this code currently does) is to just throw new
tuples on at the end. This also doesn't work because you now have 2
cells in one location, the one that was created at instatiation and the
one that you added later.

And if you fix that, then you still have to write code to return your
cells to you in the right order, otherwise your orderly Conway turns
into chaos.